+++ test/compliment/t_core.clj
@@ -105,6 +105,11 @@
                                         :ns 'compliment.t-core}))
     => (just ["class" "class?" "clojure-version" "clear-agent-errors"] :in-any-order))
 
+  (fact ":sources predicate can filter the sources to be used during completion"
+    (strip-tags (core/completions "cl" {:sources #(= :compliment.sources.ns-mappings/ns-mappings (:name %))
+                                        :ns 'compliment.t-core}))
+    => (just ["class" "class?" "clojure-version" "clear-agent-errors"] :in-any-order))
+
   (fact "empty prefix returns a list of candidates"
     (core/completions "") => (checker not-empty))